Extract OTel observability into @keryxjs/observability plugin
Move all heavy OTel SDK code out of the keryx core and into a new
first-party plugin package at packages/plugins/observability/.

Core changes:
- packages/keryx/initializers/observability.ts stripped to a no-op stub
  that establishes api.observability via declare module augmentation;
  all SDK imports and start()/stop() removed
- Remove 6 OTel SDK deps from keryx package.json (keep @opentelemetry/api
  since Connection.ts / util/tracing.ts / web.ts reference OTel types)
- Bump keryx version 0.21.0 → 0.22.0 (minor: breaking for auto-loaded obs)
- Update observability test to verify no-op behavior without the plugin

New plugin package (@keryxjs/observability@0.1.0):
- packages/plugins/observability/ — full OTel implementation
- ObservabilityPlugin extends Initializer (loadPriority=60, after stub at 50)
- Owns all SDK deps: context-async-hooks, exporter-trace-otlp-http, resources,
  sdk-metrics, sdk-trace-base, semantic-conventions, core
- index.ts exports observabilityPlugin: KeryxPlugin
- __tests__/observability.test.ts — integration tests with real Redis/Postgres
- Mirrors keryx test infra: bunfig.toml, setup.ts, .env.example

Example backend:
- example/backend/config/plugins.ts — registers observabilityPlugin
- Add @keryxjs/observability workspace dependency

Infrastructure:
- Root package.json: add workspace + lint/format/test scripts for plugin
- .github/workflows/test.yaml: add test-plugin-observability CI job (sharded)
- .github/workflows/publish.yaml: add publish-plugin-observability job;
  trigger on either package.json path change

Docs:
- docs/guide/observability.md: add Installation section with bun add + config
- docs/guide/plugins.md: add First-Party Plugins table with @keryxjs/observability
