+++ .github/workflows/publish.yaml
@@ -3,10 +3,12 @@ name: Publish
 on:
   push:
     branches: [main]
-    paths: [packages/keryx/package.json]
+    paths:
+      - packages/keryx/package.json
+      - packages/plugins/observability/package.json
 
 jobs:
-  publish:
+  publish-keryx:
     runs-on: ubuntu-latest
     permissions:
       contents: write
@@ -47,4 +49,47 @@ jobs:
           VERSION=${{ steps.version.outputs.local }}
           gh release create "v${VERSION}" \
             --title "v${VERSION}" \
+            --generate-notes
+
+  publish-plugin-observability:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+      id-token: write
+    steps:
+      - uses: actions/checkout@v4
+      - uses: oven-sh/setup-bun@v2
+      - uses: actions/setup-node@v4
+        with:
+          node-version: "22"
+          registry-url: https://registry.npmjs.org
+      - run: npm install -g npm@latest
+      - run: bun install
+
+      - name: Check if version changed
+        id: version
+        run: |
+          LOCAL=$(node -p "require('./packages/plugins/observability/package.json').version")
+          REMOTE=$(npm view @keryxjs/observability version 2>/dev/null || echo "0.0.0")
+          echo "local=$LOCAL" >> "$GITHUB_OUTPUT"
+          echo "remote=$REMOTE" >> "$GITHUB_OUTPUT"
+          if [ "$LOCAL" != "$REMOTE" ]; then
+            echo "changed=true" >> "$GITHUB_OUTPUT"
+          else
+            echo "changed=false" >> "$GITHUB_OUTPUT"
+          fi
+
+      - name: Publish to npm
+        if: steps.version.outputs.changed == 'true'
+        run: npm publish --access public --provenance
+        working-directory: packages/plugins/observability
+
+      - name: Create GitHub Release
+        if: steps.version.outputs.changed == 'true'
+        env:
+          GH_TOKEN: ${{ github.token }}
+        run: |
+          VERSION=${{ steps.version.outputs.local }}
+          gh release create "@keryxjs/observability@${VERSION}" \
+            --title "@keryxjs/observability v${VERSION}" \
             --generate-notes
