+++ .github/workflows/test.yaml
@@ -112,6 +112,46 @@ jobs:
         env:
           DATABASE_URL_TEST: postgres://postgres:postgres@localhost:5432/keryx-test
 
+  test-plugin-observability:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        shard: [1, 2]
+
+    services:
+      redis:
+        image: redis
+        ports:
+          - 6379:6379
+      postgres:
+        image: postgres
+        env:
+          POSTGRES_PASSWORD: postgres
+          POSTGRES_DB: keryx-test
+        options: >-
+          --health-cmd pg_isready
+          --health-interval 10s
+          --health-timeout 5s
+          --health-retries 5
+        ports:
+          - 5432:5432
+
+    steps:
+      - uses: actions/checkout@v4
+      - uses: oven-sh/setup-bun@v1
+      - run: bun install
+      - run: cp packages/plugins/observability/.env.example packages/plugins/observability/.env
+      - name: Run test shard ${{ matrix.shard }}/2
+        run: |
+          cd packages/plugins/observability
+          files=$(find __tests__ -name '*.test.ts' | sort)
+          shard_files=$(echo "$files" | awk "NR % 2 == (${{ matrix.shard }} - 1)")
+          echo "Running: $shard_files"
+          bun test $shard_files
+        env:
+          DATABASE_URL_TEST: postgres://postgres:postgres@localhost:5432/keryx-test
+
   test-example-frontend:
     runs-on: ubuntu-latest
     steps:
@@ -132,7 +172,7 @@ jobs:
   complete:
     runs-on: ubuntu-latest
     if: always()
-    needs: [compile, lint, test-package, test-example-backend, test-example-frontend, test-docs]
+    needs: [compile, lint, test-package, test-plugin-observability, test-example-backend, test-example-frontend, test-docs]
     steps:
       - run: |
           if [[ "${{ contains(needs.*.result, 'failure') || contains(needs.*.result, 'cancelled') }}" == "true" ]]; then
