+++ example/backend/config/plugins.ts
@@ -1,0 +1,6 @@
+import { observabilityPlugin } from "@keryxjs/observability";
+import type { KeryxPlugin } from "keryx";
+
+export default {
+  plugins: [observabilityPlugin] as KeryxPlugin[],
+};
