+++ package.json
@@ -8,6 +8,7 @@
   "description": "The fullstack TypeScript framework for MCP and APIs — write one action, serve HTTP, WebSocket, CLI, background tasks, and MCP tools.",
   "workspaces": [
     "packages/keryx",
+    "packages/plugins/observability",
     "example/backend",
     "example/frontend",
     "docs"
@@ -18,18 +19,21 @@
     "dev:example-frontend": "cd example/frontend && bun run dev",
     "tests": "bun test-package && bun test-example-backend && bun test-example-frontend",
     "test-package": "cd packages/keryx && bun test",
+    "test-plugin-observability": "cd packages/plugins/observability && bun test",
     "test-example-backend": "cd example/backend && bun test",
     "test-example-frontend": "cd example/frontend && bun test",
     "lint-package": "cd packages/keryx && bun lint",
+    "lint-plugin-observability": "cd packages/plugins/observability && bun lint",
     "lint-example-backend": "cd example/backend && bun lint",
     "lint-example-frontend": "cd example/frontend && bun lint",
     "lint-docs": "cd docs && bun lint",
-    "lint": "bun lint-package && bun lint-example-backend && bun lint-example-frontend && bun lint-docs",
+    "lint": "bun lint-package && bun lint-plugin-observability && bun lint-example-backend && bun lint-example-frontend && bun lint-docs",
     "format-package": "cd packages/keryx && bun format",
+    "format-plugin-observability": "cd packages/plugins/observability && bun format",
     "format-example-backend": "cd example/backend && bun format",
     "format-example-frontend": "cd example/frontend && bun format",
     "format-docs": "cd docs && bun format",
-    "format": "bun format-package && bun format-example-backend && bun format-example-frontend && bun format-docs",
+    "format": "bun format-package && bun format-plugin-observability && bun format-example-backend && bun format-example-frontend && bun format-docs",
     "test-docs": "cd docs && bun run test",
     "ci": "bun lint && bun tests && bun test-docs",
     "docs:dev": "cd docs && bun run dev",
