+++ packages/plugins/observability/.env.example
@@ -1,0 +1,38 @@
+PROCESS_NAME=actionhero-server
+PROCESS_NAME_TEST=test-server
+PROCESS_SHUTDOWN_TIMEOUT=30000
+
+LOG_LEVEL=info
+LOG_LEVEL_TEST=fatal
+LOG_INCLUDE_TIMESTAMPS=false
+LOG_COLORIZE=true
+
+WEB_SERVER_ENABLED=true
+WEB_SERVER_PORT=8080
+WEB_SERVER_PORT_TEST=0
+WEB_SERVER_HOST=localhost
+WEB_SERVER_API_ROUTE="/api"
+WEB_SERVER_ALLOWED_ORIGINS="http://localhost:3000"
+WEB_SERVER_ALLOWED_METHODS="GET, POST, PUT, DELETE, OPTIONS"
+
+MCP_SERVER_ENABLED=false
+
+SESSION_TTL=86400000
+SESSION_COOKIE_NAME="__session"
+
+DATABASE_URL="postgres://$USER@localhost:5432/keryx"
+DATABASE_URL_TEST="postgres://$USER@localhost:5432/keryx-test"
+DATABASE_AUTO_MIGRATE=false
+
+REDIS_URL="redis://localhost:6379/0"
+REDIS_URL_TEST="redis://localhost:6379/1"
+
+RATE_LIMIT_ENABLED=true
+RATE_LIMIT_WINDOW_MS=60000
+RATE_LIMIT_UNAUTH_LIMIT=20
+RATE_LIMIT_AUTH_LIMIT=200
+
+TASKS_ENABLED=true
+TASK_PROCESSORS=1
+TASK_TIMEOUT=5000
+TASK_TIMEOUT_TEST=100
